+++ internal/app/apptainer/remote_list.go
@@ -2,8 +2,8 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
+// Copyright (c) 2019-2023, Sylabs Inc. All rights reserved.
 // Copyright (c) 2020, Control Command Inc. All rights reserved.
-// Copyright (c) 2019, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -90,39 +90,6 @@ func RemoteList(usrConfigFile string) (err error) {
 		fmt.Fprintf(tw, listLine, n, c.Remotes[n].URI, active, sys, excl, insec)
 	}
 	tw.Flush()
-
-	if ep, err := c.GetDefault(); err == nil {
-		if err := ep.UpdateKeyserversConfig(); err == nil {
-			fmt.Println()
-			fmt.Println("Keyservers")
-			fmt.Println("==========")
-			fmt.Println()
-
-			tw = tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-			fmt.Fprintf(tw, "%s\t%s\t%s\t%s\n", "URI", "GLOBAL", "INSECURE", "ORDER")
-			order := 1
-			for _, kc := range ep.Keyservers {
-				if kc.Skip {
-					continue
-				}
-				insecure := "NO"
-				if kc.Insecure {
-					insecure = "YES"
-				}
-				fmt.Fprintf(tw, "%s\tYES\t%s\t%d", kc.URI, insecure, order)
-				if !kc.External {
-					fmt.Fprintf(tw, "*\n")
-				} else {
-					fmt.Fprintf(tw, "\n")
-				}
-				order++
-			}
-			tw.Flush()
-
-			fmt.Println()
-			fmt.Println("* Active cloud services keyserver")
-		}
-	}
 
 	if len(c.Credentials) > 0 {
 		fmt.Println()